+++ src/frontend/apps/impress/src/features/docs/doc-header/components/DocToolBox.tsx
@@ -18,11 +18,7 @@ import {
 } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
 import { useEditorStore } from '@/features/docs/doc-editor/';
-import {
-  Doc,
-  ModalRemoveDoc,
-  useCopyDocLink,
-} from '@/features/docs/doc-management';
+import { Doc, ModalRemoveDoc } from '@/features/docs/doc-management';
 import { DocShareModal } from '@/features/docs/doc-share';
 import {
   KEY_LIST_DOC_VERSIONS,
@@ -41,8 +37,6 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
   const hasAccesses = doc.nb_accesses > 1;
   const queryClient = useQueryClient();
 
-  const copyDocLink = useCopyDocLink(doc.id);
-
   const { spacingsTokens, colorsTokens } = useCunninghamTheme();
 
   const spacings = spacingsTokens();
@@ -55,24 +49,15 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
 
   const { isSmallMobile, isDesktop } = useResponsiveStore();
   const { editor } = useEditorStore();
-
   const { toast } = useToastProvider();
-  const canViewAccesses = doc.abilities.accesses_view;
 
   const options: DropdownMenuOption[] = [
     ...(isSmallMobile
       ? [
           {
-            label: canViewAccesses ? t('Share') : t('Copy link'),
-            icon: canViewAccesses ? 'group' : 'link',
-
-            callback: () => {
-              if (canViewAccesses) {
-                modalShare.open();
-                return;
-              }
-              copyDocLink();
-            },
+            label: t('Share'),
+            icon: 'group',
+            callback: modalShare.open,
           },
           {
             label: t('Export'),
@@ -165,7 +150,7 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
         $margin={{ left: 'auto' }}
         $gap={spacings['2xs']}
       >
-        {canViewAccesses && !isSmallMobile && (
+        {!isSmallMobile && (
           <>
             {!hasAccesses && (
               <Button
@@ -199,23 +184,13 @@ export const DocToolBox = ({ doc }: DocToolBoxProps) => {
                   }}
                   size={isSmallMobile ? 'small' : 'medium'}
                 >
-                  {doc.nb_accesses - 1}
+                  {doc.nb_accesses}
                 </Button>
               </Box>
             )}
           </>
         )}
-        {!canViewAccesses && !isSmallMobile && (
-          <Button
-            color="tertiary-text"
-            onClick={() => {
-              copyDocLink();
-            }}
-            size={isSmallMobile ? 'small' : 'medium'}
-          >
-            {t('Copy link')}
-          </Button>
-        )}
+
         {!isSmallMobile && (
           <Button
             color="tertiary-text"